new kernel timer

+++ src/include/types.h
@@ -13,8 +13,8 @@ typedef unsigned int mode_t;
 typedef unsigned long dev_t;
 typedef unsigned long blksize_t;
 typedef unsigned long blkcnt_t;
-typedef long clock_t;
-typedef long time_t;
+typedef unsigned long clock_t;
+typedef unsigned time_t;
 typedef int clockid_t;
 typedef unsigned long addr_t;
